+++ views/layouts/main.handlebars
@@ -11,8 +11,31 @@
     <link rel="stylesheet" href="/css/style.css">
 </head>
 <body>
+<<<<<<< HEAD
+=======
+    <nav>
+        <div class="nav-wrapper grey darken-4">
+            <a href="#" id="logo" class="brand-logo deep purple-text">play.join() Gaming Connection</a>
+            <a href="#" data-target="mobile-demo" class="sidenav-trigger"><i class="material-icons">menu</i></a>
+            <ul id="nav-mobile" class="right hide-on-med-and-down">
+                <li><a class="deep-purple" id="nav-link" href="/">Homepage</a></li>
+                <li><a class="deep-purple" id="nav-link" href="/login">Login</a></li>
+                <li><a class="deep-purple" id="nav-link" href="/profile/0">Profile</a></li>
+                <li><a class="deep-purple" id="nav-link" href="/games">Games</a></li>
+            </ul>
+        </div>
+        <h1 id="header-bar" class="hide-on-small-and-down">Recipes and Blogs</h1>
+    </nav>
+    <ul class="sidenav" id="mobile-demo">
+        <li><a id="nav-link" href="../homepage.handlebars">Homepage</a></li>
+        <li><a id="nav-link" href="../login.handlebars">Login</a></li>
+        <li><a id="nav-link" href="../profile.handlebars">Profile</a></li>
+        <li><a id="nav-link" href="#">Games</a></li>
+    </ul>
+>>>>>>> 4e419f165e36caff6004cfa30fac1f21a11e9479
     <header>
         <nav>
+<<<<<<< HEAD
             <div class="nav-wrapper black">
                 <img id="logo" src="/image/Gaming.png" alt="player.join() Gaming Connection">
                 <a href="#" data-target="mobile-demo" class="sidenav-trigger"><i class="material-icons">menu</i></a>
@@ -28,6 +51,14 @@
                     <li><a class="deep-purple" id="nav-link" href="/games">Games</a></li>
                 </ul>
             </div>
+=======
+            <a href="/">Home</a>
+            {{!-- {{#if loggedIn}} --}}
+            <a href="/profile/0">Profile</a>
+            <a id="navbar-logout" role="button">Logout</a>
+            {{!-- {{/if}} --}}
+            <a href="/login">Login</a>
+>>>>>>> 4e419f165e36caff6004cfa30fac1f21a11e9479
         </nav>
         <ul class="sidenav" id="mobile-demo">
             <li><a id="nav-link" href="../homepage.handlebars">Homepage</a></li>
