+++ views/login.handlebars
@@ -1,23 +1,20 @@
-{{#extends "main"}}
-{{#block "content"}}
-    <main class="container">
-        <section class="section">
-            <h1>Login to Your Account</h1>
-            <form method="POST" action="/api/users/login">
-                <div class="input-field">
-                    <input id="email" type="email" class="validate" name="email" required>
-                    <label for="email">Email</label>
-                </div>
-                <div class="input-field">
-                    <input id="password" type="password" class="validate" name="password" required>
-                    <label for="password">Password</label>
-                </div>
-                <button class="btn waves-effect waves-light" type="submit" name="action">Login
-                    <i class="material-icons right">send</i>
-                </button>
-            </form>
-            <p>Don't have an account? <a href="/signup">Sign up</a></p>
-            <p><a href="/forgot-password">Forgot Password?</a></p>
-        </section>
-    </main>
-{{/block}}
+<main class="container">
+    <section class="section">
+        <h1>Login to Your Account</h1>
+        <form method="POST" action="/api/users/login">
+            <div class="input-field">
+                <input id="email" type="email" class="validate" name="email" required>
+                <label for="email">Email</label>
+            </div>
+            <div class="input-field">
+                <input id="password" type="password" class="validate" name="password" required>
+                <label for="password">Password</label>
+            </div>
+            <button class="btn waves-effect waves-light" type="submit" name="action">Login
+                <i class="material-icons right">send</i>
+            </button>
+        </form>
+        <p>Don't have an account? <a href="/signup">Sign up</a></p>
+        <p><a href="/forgot-password">Forgot Password?</a></p>
+    </section>
+</main>
