+++ views/signup.handlebars
@@ -1,26 +1,23 @@
-{{#extends "main"}}
-{{#block "content"}}
-    <main class="container">
-        <section class="section">
-            <h1>Create Your Account</h1>
-            <form method="POST" action="/api/users/signup">
-                <div class="input-field">
-                    <input id="userName" type="text" class="validate" name="userName" required>
-                    <label for="userName">Username</label>
-                </div>
-                <div class="input-field">
-                    <input id="email" type="email" class="validate" name="email" required>
-                    <label for="email">Email</label>
-                </div>
-                <div class="input-field">
-                    <input id="password" type="password" class="validate" name="password" required>
-                    <label for="password">Password</label>
-                </div>
-                <button class="btn waves-effect waves-light" type="submit" name="action">Sign Up
-                    <i class="material-icons right">send</i>
-                </button>
-            </form>
-            <p>Already have an account? <a href="/login">Log in</a></p>
-        </section>
-    </main>
-{{/block}}
+<main class="container">
+    <section class="section">
+        <h1>Create Your Account</h1>
+        <form method="POST" action="/api/users/signup">
+            <div class="input-field">
+                <input id="userName" type="text" class="validate" name="userName" required>
+                <label for="userName">Username</label>
+            </div>
+            <div class="input-field">
+                <input id="email" type="email" class="validate" name="email" required>
+                <label for="email">Email</label>
+            </div>
+            <div class="input-field">
+                <input id="password" type="password" class="validate" name="password" required>
+                <label for="password">Password</label>
+            </div>
+            <button class="btn waves-effect waves-light" type="submit" name="action">Sign Up
+                <i class="material-icons right">send</i>
+            </button>
+        </form>
+        <p>Already have an account? <a href="/login">Log in</a></p>
+    </section>
+</main>
